chore: expand diagnostics.rs into existing modules
